Reorganizando arquivos do projeto
