+++ semana-10/labex/src/pages/HomePage/styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+import imgBackground from'../../assets/background.jpg'
+
+
+export const Home = styled.div`
+  background-image: url(${imgBackground});
+  background-size: cover;
+  height: 810px;
+`
+export const ContainerDiv = styled.div`
+  margin-top: 150px;
+  font-size: 14px;
+`
+export const Text = styled.h2`
+ color: white;
+`
