+++ semana-10/labex/src/pages/LoginPage/styles.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components'
+import imgBackground from'../../assets/background-space.jpg'
+
+export const LoginForm = styled.form`
+  display: grid;
+  margin-left: 640px;
+  margin-right: 640px;
+  background-color: white;
+  border-radius: 10px;
+  gap: 10px;
+`
+export const Page = styled.div`
+  background-image: url(${imgBackground});
+  background-size: cover;
+  padding-top: 50px;
+  height: 810px;
+  color: white;
+`
