+++ semana-10/labex/src/pages/TripDetailsPage/styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import imgBackground from'../../assets/background-space.jpg'
+
+export const Page = styled.div`
+  background-image: url(${imgBackground});
+  background-size: cover;
+  height: 810px;
+`
+
+export const Container = styled.div`
+  background-color: white;
+  margin-left: 100px;
+  margin-right: 100px;
+`
